fix: clear Guardian CI ratchets
