+++ desktop/src-tauri/src/managed_agents/mod.rs
@@ -14,6 +14,7 @@ pub(crate) mod effective_config;
 mod env_vars;
 pub(crate) mod git_bash;
 pub(crate) mod global_config;
+mod guardian_protection;
 mod managed_node_paths;
 mod nest;
 mod persona_avatars;
@@ -57,6 +58,7 @@ pub(crate) use global_config::{
     load_global_agent_config, resolve_effective_model_provider, save_global_agent_config,
     validate_global_config, GlobalAgentConfig,
 };
+pub use guardian_protection::*;
 pub(crate) use managed_node_paths::*;
 pub use nest::*;
 pub use personas::*;
